pyproject.toml
[tool.poetry]
name = "coding-challenge-screening"
version = "0.2.0"
description = ""
authors = ["ellamind"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
datasets = "^2.17.1"
openai = "^1.12.0"
huggingface-hub = "^0.20.3"
llama-cpp-python = {extras = ["server"], version = "^0.3.6"}
fastapi = {extras = ["all"], version = "^0.115.6"}
uvicorn = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
